OPS-1382 Renamed to package loadtests
